% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_openai.R
\name{ask_openai}
\alias{ask_openai}
\alias{extract_assistant_response}
\alias{remove_query_from_message}
\title{ask_openai}
\usage{
ask_openai(content, thread_id, assistant_id, key)

extract_assistant_response(content, roles = c("assistant", "user"))

remove_query_from_message(content)
}
\arguments{
\item{key}{OpenAI secret key}

\item{roles}{Roles to extract response from either \code{user} or \code{assistant}}

\item{project_id, dataset_id}{Bigquery path of the dataset to query to}
}
\value{
invisible.
}
\description{
Run flow that ask a question to openAI and perform query to Bigquery
}
