% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_thread.R
\name{create_thread}
\alias{create_thread}
\alias{create_thread_and_message}
\title{create_thread}
\usage{
create_thread(key)

create_thread_and_message(file_id, key, content)
}
\arguments{
\item{key}{OpenAI secret key}

\item{file_id}{id of the file in openAI account}

\item{content}{Content Text of the message}
}
\value{
invisible.
}
\description{
Create a Thread in OpenAI API
}
