% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_response.R
\name{log_response}
\alias{log_response}
\title{log_response}
\usage{
log_response(thread, file_name, log_path)
}
\arguments{
\item{thread}{the thread object retreive via \code{run_thread_openai}}

\item{file_name}{string to identify the response}

\item{log_path}{path to save the log in csv format}
}
\value{
invisible.
}
\description{
Save thread response in a csv file
}
