% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_openai.R
\name{req_openai}
\alias{req_openai}
\title{req_openai}
\usage{
req_openai(req, key, beta = TRUE, app_json = TRUE, dry_run = FALSE)
}
\arguments{
\item{req}{req object to perform the request}

\item{key}{OpenAI secret key}

\item{beta}{include beta header}

\item{app_json}{include application/json header}

\item{dry_run}{print request (to inspect) instead of performing it}
}
\value{
invisible.
}
\description{
Perform a request to openAI API, with custom headers and authentication
}
