% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message}
\alias{send_message}
\alias{list_messages}
\title{send_message}
\usage{
send_message(thread_id, content, key, file_id = NULL)

list_messages(thread_id, key, after_mssg = NULL, before_mssg = NULL)
}
\arguments{
\item{thread_id}{ID of the thread}

\item{content}{Content Text of the message}

\item{key}{OpenAI secret key}

\item{file_id}{id of the file in openAI account}

\item{after_mssg, before_mssg}{parameters to retreive messages before or after a specific message}
}
\value{
invisible.
}
\description{
Send a user message to OpenAI thread
}
