% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\alias{remove_file}
\alias{retrieve_file_content}
\title{upload_file}
\usage{
upload_file(df, key, file_path, ...)

remove_file(file_id)

retrieve_file_content(file_id, key)
}
\arguments{
\item{df}{dataframe to upload as csv}

\item{key}{OpenAI secret key}

\item{file_path}{path of the local csv file}

\item{...}{other arguments to pass to req_openai}

\item{file_id}{id of the file in openAI account}
}
\value{
invisible.
}
\description{
Perform a request to openAI API regarding files.
}
