% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vkchat-package.R
\docType{package}
\name{vkchat-package}
\alias{vkchat}
\alias{vkchat-package}
\title{vkchat: Easy communication with OpenAI API to enable chat-based applications}
\description{
Wrappers to facilitate use of OpenAI API, currently two approach are developed: `Insights` to enable knowledge creation based on specific dataframes `Idiomatic` chat-with-your-data type of application.
}
\author{
\strong{Maintainer}: Rodrigo Serrano \email{roserrano@vikua.com}

}
\keyword{internal}
