% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_openai.R
\name{ask_openai}
\alias{ask_openai}
\title{ask_openai}
\usage{
ask_openai(content, thread_id, assistant_id, key, project_id, dataset_id)
}
\arguments{
\item{key}{OpenAI secret key}

\item{project_id, dataset_id}{Bigquery path of the dataset to query to}

\item{roles}{Roles to extract response from either \code{user} or \code{assistant}}
}
\value{
invisible.
}
\description{
Run workflow that ask a question to openAI and perform query to Bigquery
}
